Convert to ruby gem
