+++ .gitignore
@@ -1,11 +1,47 @@
 buildArtifacts/
 **/.DS_Store
 .DS_Store
-.bundle
-vendor
 .idea
 *.xcworkspace
 !default.xcworkspace
 xcuserdata
 SampleApp/AutomatorSampleApp/build
 SampleApp/AutomatorSampleApp/Pods
+
+## For gem
+
+*.gem
+*.rbc
+/.config
+/coverage/
+/InstalledFiles
+/pkg/
+/spec/reports/
+/test/tmp/
+/test/version_tmp/
+/tmp/
+
+## Specific to RubyMotion:
+.dat*
+.repl_history
+build/
+
+## Documentation cache and generated files:
+/.yardoc/
+/_yardoc/
+/doc/
+/rdoc/
+
+## Environment normalisation:
+/.bundle/
+/vendor/bundle
+/lib/bundler/man/
+
+# for a library or gem, you might want to ignore these files since the code is
+# intended to run in multiple environments; otherwise, check them in:
+Gemfile.lock
+# .ruby-version
+# .ruby-gemset
+
+# unless supporting rvm < 1.11.0 or doing something fancy, ignore this:
+.rvmrc
