+++ .travis.yml
@@ -1,2 +1,3 @@
 language: objective-c
+install: bundle install --jobs=3 --retry=3 --gemfile SampleApp/Gemfile
 script: ruby SampleApp/smokeTests.rb
