+++ SampleApp/Gemfile
@@ -1,0 +1,2 @@
+source 'https://rubygems.org'
+gem 'illuminator', :path => '../'
