+++ gem/lib/illuminator/build-artifacts.rb
