+++ gem/lib/illuminator/listeners/pretty-output.rb
@@ -1,5 +1,5 @@
 
-require File.join(File.expand_path(File.dirname(__FILE__)), 'InstrumentsListener.rb')
+require_relative './instruments-listener'
 
 class PrettyOutput < InstrumentsListener
