+++ gem/lib/illuminator/listeners/saltinel-agent.rb
@@ -1,5 +1,5 @@
 
-require File.join(File.expand_path(File.dirname(__FILE__)), 'SaltinelListener.rb')
+require_relative './saltinel-listener'
 
 module SaltinelAgentEventSink
