+++ gem/lib/illuminator/listeners/test-listener.rb
@@ -1,6 +1,6 @@
 require 'date'
 
-require File.join(File.expand_path(File.dirname(__FILE__)), 'InstrumentsListener.rb')
+require_relative './instruments-listener'
 
 module TestListenerEventSink
