+++ gem/lib/illuminator/resources/IlluminatorGeneratedEnvironment.erb
