+++ gem/lib/illuminator/resources/IlluminatorGeneratedRunnerForInstruments.erb
