+++ gem/lib/illuminator/version.rb
@@ -1,0 +1,3 @@
+module Illuminator
+  VERSION = "0.1.0"
+end
