+++ gem/lib/illuminator/xcode-builder.rb
@@ -1,9 +1,7 @@
-require 'bundler/setup'
-require 'rubygems'
 require 'colorize'
 
-require File.join(File.expand_path(File.dirname(__FILE__)), 'BuildArtifacts.rb')
-require File.join(File.expand_path(File.dirname(__FILE__)), 'HostUtils.rb')
+require_relative './build-artifacts'
+require_relative './host-utils'
 
 class XcodeBuilder
   attr_accessor :project
