--- gemfile
@@ -1,5 +1,0 @@
-source 'https://rubygems.org'
-gem 'dnssd'
-gem 'colorize'
-gem 'json'
-gem 'xcpretty'
